extra field verification added
